AI_Engineer/Backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-genai>=0.2.0
python-multipart>=0.0.6
pillow>=10.0.0
pydantic>=2.5.0
python-dotenv
requests
